+++ src/shared/stats.rs
@@ -5,7 +5,10 @@
 // <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
-use super::super::Config;
+
+
+// Internal Dependencies ------------------------------------------------------
+use ::Config;
 
 /// A structure containing stats data average of the course of one second.
 #[derive(Debug, PartialEq, Copy, Clone)]
@@ -43,7 +46,7 @@ impl Default for Stats {
 pub struct StatsCollector {
 
     /// Internal tick value
-    tick: u32,
+    tick: u64,
 
     /// The collectors's configuration
     config: Config,
